prometheus/tests/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prometheus-testing-framework"
version = "0.1.0"
description = "Comprehensive Prometheus installation and testing framework"
readme = "../README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Prometheus Testing Team"}
]
keywords = ["prometheus", "testing", "monitoring", "kubernetes"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "hypothesis>=6.92.0",
    "httpx>=0.25.0",
    "kubernetes>=28.1.0",
    "pyyaml>=6.0.1",
    "jsonschema>=4.20.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "prometheus-client>=0.19.0",
    "jinja2>=3.1.2",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.7.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "flake8>=6.1.0",
    "pytest-cov>=4.1.0",
]
load = [
    "locust>=2.20.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
]

[project.scripts]
prom-test = "cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
